refactor Rt page
